Added citation reminder in the head of pyunicorn files.

+++ pyunicorn/timeseries/_ext/numerics.pyx
@@ -4,6 +4,15 @@
 # Copyright (C) 2008--2019 Jonathan F. Donges and pyunicorn authors
 # URL: <http://www.pik-potsdam.de/members/donges/software>
 # License: BSD (3-clause)
+#
+# Please acknowledge and cite the use of this software and its authors
+# when results are used in publications or published elsewhere.
+#
+# You can use the following reference:
+# J.F. Donges, J. Heitzig, B. Beronov, M. Wiedermann, J. Runge, Q.-Y. Feng,
+# L. Tupikina, V. Stolbova, R.V. Donner, N. Marwan, H.A. Dijkstra,
+# and J. Kurths, "Unified functional network and nonlinear time series analysis
+# for complex systems science: The pyunicorn package"
 
 cimport cython
 from cpython cimport bool
@@ -48,9 +57,9 @@ cdef extern from "time.h":
 
 cdef extern from "src_numerics.c":
     void _manhattan_distance_matrix_fast(int ntime_x, int ntime_y, int dim,
-        double *x_embedded, double *y_embedded, float *distance)
+        double *x_embedded, double *y_embedded, double *distance)
     void _euclidean_distance_matrix_fast(int ntime_x, int ntime_y, int dim,
-        double *x_embedded, double *y_embedded, float *distance)
+        double *x_embedded, double *y_embedded, double *distance)
     void _supremum_distance_matrix_fast(int ntime_x, int ntime_y, int dim,
         float *x_embedded, float *y_embedded, float *distance)
     void _test_pearson_correlation_fast(double *original_data,
@@ -68,14 +77,14 @@ def _manhattan_distance_matrix_crp(
     np.ndarray[double, ndim=2, mode='c'] x_embedded not None,
     np.ndarray[double, ndim=2, mode='c'] y_embedded not None):
 
-    cdef np.ndarray[float, ndim=2, mode='c'] distance = \
-        np.zeros((ntime_x, ntime_y), dtype="float32")
+    cdef np.ndarray[double, ndim=2, mode='c'] distance = \
+        np.zeros((ntime_x, ntime_y), dtype="double")
 
     _manhattan_distance_matrix_fast(
         ntime_x, ntime_y, dim,
         <double*> np.PyArray_DATA(x_embedded),
         <double*> np.PyArray_DATA(y_embedded),
-        <float*> np.PyArray_DATA(distance))
+        <double*> np.PyArray_DATA(distance))
 
     return distance
 
@@ -85,14 +94,14 @@ def _euclidean_distance_matrix_crp(
     np.ndarray[double, ndim=2, mode='c'] x_embedded not None,
     np.ndarray[double, ndim=2, mode='c'] y_embedded not None):
 
-    cdef np.ndarray[float, ndim=2, mode='c'] distance = \
-        np.zeros((ntime_x, ntime_y), dtype="float32")
+    cdef np.ndarray[double, ndim=2, mode='c'] distance = \
+        np.zeros((ntime_x, ntime_y), dtype="double")
 
     _euclidean_distance_matrix_fast(
         ntime_x, ntime_y, dim,
         <double*> np.PyArray_DATA(x_embedded),
         <double*> np.PyArray_DATA(y_embedded),
-        <float*> np.PyArray_DATA(distance))
+        <double*> np.PyArray_DATA(distance))
 
     return distance
